+++ packages/open-brain-cli/bin/brain.js
@@ -2,7 +2,7 @@
 import { Command } from 'commander';
 import queryCmd from '../commands/query.js';
 import relationsCmd from '../commands/find-relations.js';
-import wikiCmd from '../commands/wiki.js';
+import smokeCmd from '../commands/smoke-test.js';
 
 const program = new Command();
 
@@ -14,5 +14,6 @@ program
 program.addCommand(queryCmd);
 program.addCommand(relationsCmd);
 program.addCommand(wikiCmd);
+program.addCommand(smokeCmd);
 
 program.parse();